Tests/LibWeb/Text/expected/css/CSSRule-cssText-preserve-color-names.txt

#lowercase { stop-color: cyan; }
#uppercase { background-color: magenta; }
#mixed-case { border-color: yellow; }
#hex-value-not-preserved { color: rgb(170, 187, 204); }
